pet-platform/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
PyMySQL==1.1.1
Werkzeug==3.0.3
cryptography==42.0.8
python-dotenv==1.0.1
